Better types for app.run
Using Unit on `.run` makes use `.map(run)`  which is all Unit. `.map` should not be use for Unit. We could use `.foreach` alternatively, but then we loose the ability to do `.terminateActorSystem`. Changing to `Done`, we force a return type of `Option[Done]` and then we add termiantion through extension methods